Changed: #1206 Settings are located in a centralized place(plugin manager). Update core and example plugin.

+++ code/nel/tools/3d/object_viewer_qt/src/camera_control.cpp
@@ -183,7 +183,7 @@ CCameraControl::CCameraControl(QWidget *parent)
 
 CCameraControl::~CCameraControl()
 {
-	for(int i = 0; i < _cameraList.size(); ++i)
+	for(size_t i = 0; i < _cameraList.size(); ++i)
 		delete _cameraList[i];
 	_cameraList.clear();
 }
